Refactor ndt-server into a small main
Creates a directory ndt to service legacy clients.
